Moving tests into module folders
