+++ modules/core/data/README.md
@@ -1,0 +1,3 @@
+# Attributions for Sample Files
+
+* `image1-preview.*` - copied from `binary-gltf-utils` under MIT license: Copyright (c) 2016-17 Karl Cheng
