+++ modules/core/test/binary-utils/flatten.spec.js
