+++ modules/core/test/binary-utils/text-encoder.spec.js
