+++ modules/core/test/data-uri-utils/decode-data-uri.spec.js
