+++ modules/core/test/file-utils/load-image.spec.js
