+++ modules/core/test/image-utils/get-image-size.spec.js
@@ -8,27 +8,27 @@ import path from 'path';
 import {getImageSize} from '@loaders.gl/core';
 import {loadBinaryFile} from '@loaders.gl/core-node';
 
-const TEST_DATA_DIR = path.resolve(__dirname, '../../../data');
+const TEST_DATA_DIR = path.resolve(__dirname, '../../data');
 
 const PNG_BINARY =
   loadBinaryFile(path.resolve(TEST_DATA_DIR, 'images/img1-preview.png')) ||
-  require('test-data/images/img1-preview.png');
+  require('../../data/img1-preview.png');
 
 const JPEG_BINARY =
   loadBinaryFile(path.resolve(TEST_DATA_DIR, 'images/img1-preview.jpeg')) ||
-  require('test-data/images/img1-preview.jpeg');
+  require('../../data/img1-preview.jpeg');
 
 const GIF_BINARY =
   loadBinaryFile(path.resolve(TEST_DATA_DIR, 'images/img1-preview.gif')) ||
-  require('test-data/images/img1-preview.gif');
+  require('../../data/img1-preview.gif');
 
 const BMP_BINARY =
   loadBinaryFile(path.resolve(TEST_DATA_DIR, 'images/img1-preview.bmp')) ||
-  require('test-data/images/img1-preview.bmp');
+  require('../../data/img1-preview.bmp');
 
 const TIFF_BINARY =
   loadBinaryFile(path.resolve(TEST_DATA_DIR, 'images/img1-preview.tiff')) ||
-  require('test-data/images/img1-preview.png');
+  require('../../data/img1-preview.png');
 
 const TEST_FILES = [
   ['png', PNG_BINARY],
