+++ modules/core/test/image-utils/image-loaders.spec.js
