+++ modules/core/test/index.js
