+++ modules/core/test/parser-utils/auto-parse.spec.js
